enhance script

+++ scripts/createRoutes.js
@@ -29,19 +29,25 @@ function getFileRoutes(routes) {
 
 const repoName = process.argv[2] ?? 'react-repo-name'
 const reactRepo = resolve(import.meta.dirname, '..', '..', repoName)
-const reactRouterRoutes = await getRoutes(repoName, 'npx react-router routes --json')
-const svelteRoutes = getFileRoutes(reactRouterRoutes[0].children[0].children)
-
-for (const svelteRoute of svelteRoutes) {
-  const routePath = resolve(import.meta.dirname, '..', 'src', 'routes', svelteRoute.sveltePath)
-  if (!existsSync(routePath)) {
-    mkdirSync(routePath, {recursive: true})
+if (existsSync(reactRepo)) {
+  const reactRouterRoutes = await getRoutes(repoName, 'npx react-router routes --json')
+  const layoutRoutes = reactRouterRoutes[0].children.filter(({file}) => file.endsWith('.tsx'))
+  console.log(`found main route with ${layoutRoutes[0].children.length} entries`)
+  const svelteRoutes = getFileRoutes(layoutRoutes[0].children)
+  for (const svelteRoute of svelteRoutes) {
+    const routePath = resolve(import.meta.dirname, '..', 'src', 'routes', svelteRoute.sveltePath)
+    if (!existsSync(routePath)) {
+      mkdirSync(routePath, {recursive: true})
+    }
+    const reactCode = await readFile(resolve(reactRepo, 'app', svelteRoute.tsxFile), {
+      encoding: 'utf-8'
+    })
+    const serverPart = getServerPart(reactCode)
+    const clientPart = getClientPart(reactCode)
+    console.log(`writing svelte route ${routePath}`)
+    await writeFile(format({name: '+page', ext: 'server.ts', dir: routePath}), serverPart)
+    await writeFile(format({name: '+page', ext: 'svelte', dir: routePath}), clientPart)
   }
-  const reactCode = await readFile(resolve(reactRepo, 'app', svelteRoute.tsxFile), {
-    encoding: 'utf-8'
-  })
-  const serverPart = getServerPart(reactCode)
-  const clientPart = getClientPart(reactCode)
-  await writeFile(format({name: '+page', ext: 'server.ts', dir: routePath}), serverPart)
-  await writeFile(format({name: '+page', ext: 'svelte', dir: routePath}), clientPart)
+} else {
+  console.log(`react repo ${reactRepo} not found`)
 }
